+++ components/camel-corda/pom.xml
@@ -61,30 +61,12 @@
         </dependency>
     </dependencies>
 
-    <profiles>
-        <profile>
-            <id>corda-integration</id>
-            <build>
-                <plugins>
-                    <plugin>
-                        <artifactId>maven-surefire-plugin</artifactId>
-                        <configuration>
-                            <forkedProcessTimeoutInSeconds>300</forkedProcessTimeoutInSeconds>
-                            <childDelegation>false</childDelegation>
-                            <useFile>true</useFile>
-                            <failIfNoTests>false</failIfNoTests>
-                            <runOrder>alphabetical</runOrder>
-                            <includes>
-                                <include>**/*IntegrationTest.java</include>
-                            </includes>
-                            <excludes>
-                                <exclude>**/*XXXTest.java</exclude>
-                            </excludes>
-                        </configuration>
-                    </plugin>
-                </plugins>
-            </build>
-        </profile>
-    </profiles>
-
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-failsafe-plugin</artifactId>
+            </plugin>
+        </plugins>
+    </build>
 </project>
